Support dictionary data structure
According to the [Swagger specification]
(http://swagger.io/specification/#model-with-map-dictionary-properties-88)
a dict is defined by the following json:

```
{
  "type": "object",
  "additionalProperties": {
    "type": "string"
  }
}
```
This PR checks if the field `additionalProperties` is defined and
`properties` is undefined. If so, it interprets it as a dict, otherwise
as an object as usual. It will use the nested field `type` as the second
type of the Elm dictionary. The above example would result in a
dictionary `Dict String String`.

Related to issue #10 and [Alertmanager
PR](https://github.com/prometheus/alertmanager/pull/698)